requirements.txt
langgraph>=0.2.50
langchain-core>=0.3.0
openai>=1.40.0
mcp>=1.2.0
httpx>=0.27.0
pydantic>=2.7.0
python-dotenv>=1.0.1
PyYAML>=6.0.1
jsonschema>=4.23.0
requests>=2.32.0
pytest>=8.0.0
